Removed Non mastery factors section
The non mastery section duplicated the "rationale" page, and drew a line between mastery standards and non mastery, instead of content and non content standards, which is not correct. We still may need to explain that the success standards should be a part of the grade, but they are still something that students should try to obtain with mastery.