Students receive NU to indicate course extension
It should remain `NU` - see below the academic policy for awarding grades of `NX`. Our course extension process uses `NU` to create a parallel, but different, process under which students are permitted to submit work up to the end of Term 7

> Schools may award a grade of incomplete (‘NX’) if a student has a documented, extreme
extenuating circumstance that prevents him/her from completing the course in its established timeframe
(e.g., surgery, death in the family). A student who receives an incomplete must successfully complete
remaining course requirements by the end of the term following the termination of the course in order
to receive a final grade and credit, as applicable. ‘NX’ does not have a pass/fail or a numeric equivalent.