Correct Syntax
Markdown links require no space between bracket sets, like this:
[link text](http://…)

Headers are shown with a number of hashes, `#`, the number indicating
the level of the header with more hashes meaning a “smaller” header.
There is always a white space between the hash and the title itself,
i.e,

### Header Level 3

Style for this policy is to include one blank line immediately before,
and immediately after, any header.